// dashboard/src/client/models/hub-workflow-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { HubPageMeta } from './hub-page-meta';
// May contain unused imports in some cases
// @ts-ignore
import type { HubWorkflowSummary } from './hub-workflow-summary';

export interface HubWorkflowListResponse {
    'items': Array<HubWorkflowSummary>;
    'meta': HubPageMeta;
}