// dashboard/src/client/models/hub-workflow-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { HubVisibility } from './hub-visibility';

export interface HubWorkflowSummary {
    'id': string;
    'name': string;
    'summary'?: string | null;
    'description'?: string | null;
    'tags'?: Array<string> | null;
    'ownerId'?: string | null;
    'ownerName'?: string | null;
    'updatedAt'?: string | null;
    'latestVersion'?: string | null;
    'previewImage'?: string | null;
    'visibility'?: HubVisibility;
}